+++ src/dependency-manager/spec/component-spec.ts
@@ -3,7 +3,7 @@ import { Exclude, Transform } from 'class-transformer';
 import { Allow, IsOptional, IsString, ValidateNested } from 'class-validator';
 import { JSONSchema } from 'class-validator-jsonschema';
 import { Dictionary } from '../utils/dictionary';
-import { ServiceSpec } from './service-spec';
+import { IngressSpec, ServiceSpec } from './service-spec';
 import { TaskSpec } from './task-spec';
 import { transformObject } from './transform/common-transform';
 import { AnyOf, ArrayOf, ExpressionOr, ExpressionOrString } from './utils/json-schema-annotations';
@@ -24,45 +24,6 @@ export interface ComponentInstanceMetadata {
   }
 
   deprecated_interfaces_map: Dictionary<string | undefined>;
-}
-
-@JSONSchema({
-  description: 'An ingress exposes an interface to external network traffic through an architect-deployed gateway.',
-})
-export class IngressSpec {
-  @IsOptional()
-  @JSONSchema({
-    type: 'boolean',
-    description: 'Marks the interface as an ingress.',
-  })
-  enabled?: boolean;
-
-  @IsOptional()
-  @JSONSchema({
-    ...ExpressionOr({ type: 'string', pattern: Slugs.ComponentSubdomainValidator.source }),
-    description: 'The subdomain that will be used if the interface is exposed externally',
-    errorMessage: Slugs.ComponentSubdomainDescription,
-  })
-  subdomain?: string;
-
-  @IsOptional()
-  @JSONSchema({
-    ...ExpressionOr({ type: 'string', pattern: '^\\/.*$' }),
-    description: 'The path of the interface used for path based routing',
-  })
-  path?: string;
-
-  @IsOptional()
-  @JSONSchema({
-    ...ExpressionOr({
-      type: 'array',
-      items: {
-        anyOf: [{ type: 'string', format: 'cidrv4' }, { type: 'string', pattern: '\\${{\\s*secrets\\.[\\w-]+\\s*}}' }],
-      },
-    }),
-    description: 'IP addresses that are allowed to access the interface',
-  })
-  ip_whitelist?: string[];
 }
 
 @JSONSchema({
