+++ src/dependency-manager/spec/service-spec.ts
@@ -5,11 +5,49 @@ import { JSONSchema } from 'class-validator-jsonschema';
 import { DeepPartial } from '../../common/utils/types';
 import { Dictionary } from '../utils/dictionary';
 import { LivenessProbeSpec, VolumeSpec } from './common-spec';
-import { IngressSpec } from './component-spec';
 import { ResourceSpec } from './resource-spec';
 import { transformObject } from './transform/common-transform';
 import { AnyOf, ExclusiveOr, ExpressionOr, ExpressionOrString, RequiredOr } from './utils/json-schema-annotations';
 import { Slugs } from './utils/slugs';
+
+@JSONSchema({
+  description: 'An ingress exposes an interface to external network traffic through an architect-deployed gateway.',
+})
+export class IngressSpec {
+  @IsOptional()
+  @JSONSchema({
+    type: 'boolean',
+    description: 'Marks the interface as an ingress.',
+  })
+  enabled?: boolean;
+
+  @IsOptional()
+  @JSONSchema({
+    ...ExpressionOr({ type: 'string', pattern: Slugs.ComponentSubdomainValidator.source }),
+    description: 'The subdomain that will be used if the interface is exposed externally',
+    errorMessage: Slugs.ComponentSubdomainDescription,
+  })
+  subdomain?: string;
+
+  @IsOptional()
+  @JSONSchema({
+    ...ExpressionOr({ type: 'string', pattern: '^\\/.*$' }),
+    description: 'The path of the interface used for path based routing',
+  })
+  path?: string;
+
+  @IsOptional()
+  @JSONSchema({
+    ...ExpressionOr({
+      type: 'array',
+      items: {
+        anyOf: [{ type: 'string', format: 'cidrv4' }, { type: 'string', pattern: '\\${{\\s*secrets\\.[\\w-]+\\s*}}' }],
+      },
+    }),
+    description: 'IP addresses that are allowed to access the interface',
+  })
+  ip_whitelist?: string[];
+}
 
 @JSONSchema({
   ...RequiredOr('cpu', 'memory'),
